Updated Dockerfile to add an entrypoint.

+++ docker-entrypoint.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
-cp -r /usr/src/facturascripts/* /var/www/html/; \
-chmod -R o+w /var/www/html
+
+FILE=/var/www/html/.htaccess
+if [ ! -f "$FILE" ]; then
+	cp -r /usr/src/facturascripts/* /var/www/html/; \
+	chmod -R o+w /var/www/html
+fi
+
 exec "$@"
